+++ sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2026-05-02T22:18:14+0000
+// 2026-05-28T00:04:50+0000
 (function () {
 "use strict";